+++ packages/plugins/adapters/financial/akshare/index.ts
@@ -1,0 +1,1 @@
+export * from './akshare-financial-plugin.ts'